moved base.py and distributions.py into utils.py, updated other files accordingly.
