+++ README.md
@@ -4,24 +4,7 @@ uniMASK is a generalization of BERT models with flexible abstractions for perfor
 sequences. Masking and prediction can occur both on the token level (as in traditional transformer), or even on 
 subportions of tokens.
 
-File structure:
-- `data/`: where rollouts (`datasets`) and trained models (`transformer_runs`) are stored. 
-- `envs/`: data-handling and evaluation for each supported environment. Currently
-- `scripts/`: 
-- `base.py`:
-- `batches.py`: has all data pipeline processing classes (`FactorSeq, TokenSeq, FullTokenSeq, Batch, SubBatch`)
-- `distributions.py`
-- `sequences.py`:
-- `trainer.py`:
-- `transformer.py`: contains the transformer model class itself
-- `transformer_train.py`: interface and config setting for training a transformer, through `Trainer` class
-- `utils.py`
-- `transformer_eval.py`: interface for getting predictions from transformer (currently empty)
-- 
-
 # Getting Started
-
-###	Installation
 To install uniMASK, run:
 
 ```bash
@@ -45,9 +28,22 @@ To verify that the installation was successful, run `pytest`.
 
 # Reproducing results from the paper
 ### Minigrid heatmap (figure 7)
-1. Reproduce the runs found in `minigrid_repro.sh`.
-We recommend running these in parallel, for example by running the script with GNU Parallel (an example can be found at
-first line of that file.)
+Note: Reproducing all runs can a long time. We recommend parallelizing runs.
+In each script, the first line (comment) contains an example of how to use GNU Parallel towards this end.
+
+1. Run the commands found in `minigrid_repro.sh`.
 2. Fine-tune the pre-trained models generated in the previous step by running the commands in `minigrid_ft_repro.sh`.
-You may want to parallelize these runs as above. 
-3. Reproduce the heatmaps by running `minigrid_heatmap.sh`.
+You may want to parallelize these runs as above.
+3. Generate the heatmaps from these runs by running `minigrid_heatmap.sh` (no parallelization here).
+
+# File structure
+- `data/`: where rollouts (`datasets`) and trained models (`transformer_runs`) are stored. 
+- `envs/`: data-handling and evaluation for each supported environment. Currently
+- `scripts/`:
+- `batches.py`: has all data pipeline processing classes (`FactorSeq, TokenSeq, FullTokenSeq, Batch, SubBatch`)
+- `sequences.py`:
+- `trainer.py`:
+- `transformer.py`: contains the transformer model class itself.
+- `transformer_train.py`: interface and config setting for training a transformer, through `Trainer` class.
+- `utils.py`: misc utilities, namely math functions, gpu handling, profiling, etc.
+- `transformer_eval.py`: interface for getting predictions from transformer (currently empty).
